Move icons to their files
