+++ packages/zendesk-client/src/index.ts
@@ -26,5 +26,7 @@ export type * from './types';
 export { zendeskMessageConverter } from './zendesk-message-converter';
 export { useManagedZendeskChat } from './use-managed-zendesk-chat';
 export { useGetZendeskConversations } from './use-managed-zendesk-chat';
-export { CSATForm, ThumbsUpIcon, ThumbsDownIcon } from './components/csat-form';
+export { CSATForm } from './components/csat-form';
 export type { CSATFormProps } from './components/csat-form';
+export { ThumbsUpIcon } from './components/thumbs-up-icon';
+export { ThumbsDownIcon } from './components/thumbs-down-icon';
